requirements.txt
poetry
msal
python-dotenv
flask
flask_swagger_ui
gunicorn